Fix range operator subtraction overflow on large ranges
Expression like `--80-33333333333333333333..338` produced start/end
values where `end - start` overflowed i64. Use i128 arithmetic for
the count computation to handle the full i64 range safely.

Found by cargo-fuzz (fuzz_eval, 1-hour campaign).

+++ crates/gnata-rs/src/evaluator/mod.rs
@@ -2143,7 +2143,7 @@ fn apply_range(left: &Value, right: &Value) -> JsonataResult {
     if start > end {
         return Ok(Value::Undefined);
     }
-    let count = (end - start + 1) as usize;
+    let count = (i128::from(end) - i128::from(start) + 1) as usize;
     if count > 10_000_000 {
         return Err(JsonataError::new("D2014", "range operator too large"));
     }
